# api/openapi.yaml
openapi: 3.1.0
info:
  title: SCNT Tracking System API
  version: 1.0.0
  summary: Zebrafish somatic cell nuclear transfer embryo and clone-fish tracking
  description: |
    Contract between the static SPA front end and the Go back end.

    This document is the single source of truth for the interface
    (SRS v1.0 section 8, API-02). Generate both sides from it:

      * server stubs — generated directly from this OpenAPI 3.1 document
      * browser client — `openapi-typescript api/openapi.yaml -o frontend/src/api/schema.d.ts`

    ### Conventions

    * All timestamps are RFC 3339 **with an explicit offset** (CI-04).
      The server stores UTC; the client sends and displays `Asia/Bangkok`.
    * All hour values are decimal hours with 4 decimal places (DR-05).
    * Writes carry `X-Operator-Id` and `X-Device-Id` (API-03) because the
      system has no login (CON-01) and still has to record who did what.
    * Every mutation also accepts `X-Idempotency-Key`. Clients persist this
      stable key with the queued write; retries return the original result.
    * Bulk write endpoints are idempotent on `clientUuid` (BR-18). Re-sending
      is always safe and is the normal case, not an error — the offline write
      queue retries freely.
    * JSON fields use `camelCase`; SQL columns and tabular exports use
      `snake_case`.
    * The server computes authoritative `hpaActual`, `hpaExpected`, and
      `deviationH`. The client may calculate a live preview only (DC-05).
...
servers:
...
paths:
...
  /due-checkpoints:
    get:
      tags: [Observations]
      summary: What needs looking at right now
      description: |
        The home screen (FR-401). Returns every (injection lot × checkpoint) that
        is due but not yet recorded, most overdue first, computed per BR-22:
        `dueAt = activatedAt + expectedHpa`.

        Also returns the next not-yet-due checkpoint per lot so the technician
        can see what is coming, and the count of embryos waiting to be promoted
        (FR-404).
      operationId: listDueCheckpoints
      parameters:
        - { name: siteId, in: query, schema: { type: string, format: uuid } }
        - { name: operatorId, in: query, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Work queue
          content:
            application/json:
              schema:
                type: object
                required: [overdue, upcoming, pendingPromotionCount]
                properties:
                  overdue:
                    type: array
                    items: { $ref: '#/components/schemas/DueCheckpoint' }
                  upcoming:
                    type: array
                    items: { $ref: '#/components/schemas/DueCheckpoint' }
                  pendingPromotionCount: { type: integer, minimum: 0 }
  /injection-lots/{id}/checkpoints/{stageCode}:
    parameters:
      - $ref: '#/components/parameters/PathId'
      - name: stageCode
        in: path
        required: true
        schema: { type: string }
        examples: { default: { value: stage_09_256C } }
    get:
      tags: [Observations]
      summary: Everything the checkpoint entry screen needs
      description: |
        Returns only embryos with no exit event (FR-405), each with the condition
        carried forward from its previous observation so the screen can default
        sensibly (FR-409), plus the reference time for this checkpoint so the
        client can show the deviation live as the technician works (FR-412).
      operationId: getCheckpointEntry
      responses:
        '200':
          description: Entry screen payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckpointEntry' }
        '404': { $ref: '#/components/responses/NotFound' }
  /observations/embryo:
    post:
      tags: [Observations]
      summary: Record embryo observations (bulk, idempotent)
      description: |
        The main write path of the system. Send the whole checkpoint in one call.

        **Idempotent on `clientUuid`** (BR-18): a repeat of a `clientUuid` the
        server already has returns `duplicate` with the stored row and HTTP 200 —
        never an error, never a second row. The offline queue depends on this;
        without it a request that succeeded but whose response was lost would
        double-count and corrupt the survival numbers.

        Per item the server computes `hpaActual` (BR-02), reads `hpaExpected`
        from the profile pinned to the batch and **freezes it onto the row**
        (BR-03), and derives `deviationH` (BR-04).

        Recording `DEAD` or `DEGENERATED` also writes the embryo's exit event.
        Marking an embryo `ALIVE` after it already has one violates monotonic
        survival (BR-07) and is rejected with `409` unless `overrideReason` is
        supplied.

        Partial success is normal: each item reports its own status, so one bad
        row never discards the rest of the checkpoint.
      operationId: createEmbryoObservations
      parameters: [ $ref: '#/components/parameters/OperatorId', $ref: '#/components/parameters/DeviceId', $ref: '#/components/parameters/IdempotencyKey' ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [observations]
              properties:
                observations:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items: { $ref: '#/components/schemas/EmbryoObservationInput' }
            examples:
              checkpoint:
                summary: One checkpoint, one embryo alive and one dead
                value:
                  observations:
                    - clientUuid: 0192f3a1-7c4e-7000-8000-1a2b3c4d5e6f
                      embryoId: 0192f0aa-0000-7000-8000-000000000001
                      stageCode: stage_09_256C
                      observedAt: '2026-08-20T12:38:00+07:00'
                      outcome: ALIVE
                      condition: NORMAL
                    - clientUuid: 0192f3a1-7c4e-7000-8000-1a2b3c4d5e70
                      embryoId: 0192f0aa-0000-7000-8000-000000000002
                      stageCode: stage_09_256C
                      observedAt: '2026-08-20T12:38:00+07:00'
                      outcome: DEAD
                      condition: UNDETERMINED
      responses:
        '200':
          description: Per-item results
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/EmbryoObservationResult' }
              examples:
                mixed:
                  summary: One created, one already seen
                  value:
                    results:
                      - clientUuid: 0192f3a1-7c4e-7000-8000-1a2b3c4d5e6f
                        id: 0192f3b0-0000-7000-8000-000000000001
                        status: created
                        hpaActual: 2.6333
                        hpaExpected: 2.5
                        deviationH: 0.1333
                        deviationLabel: ช้ากว่าสากล 8 นาที
                        exitRecorded: false
                      - clientUuid: 0192f3a1-7c4e-7000-8000-1a2b3c4d5e70
                        id: 0192f3b1-0000-7000-8000-000000000002
                        status: duplicate
                        hpaActual: 2.6333
                        hpaExpected: 2.5
                        deviationH: 0.1333
                        deviationLabel: ช้ากว่าสากล 8 นาที
                        exitRecorded: true
  /observations/embryo/{id}:
    parameters: [ $ref: '#/components/parameters/PathId' ]
    patch:
      tags: [Observations]
      summary: Correct a mis-entered embryo observation
      description: |
        Corrections are recorded, not hidden (FR-1101, FR-1102). Editing the
        outcome recalculates the embryo's exit event; editing the condition can
        clear or move `firstAbnormal` (BR-12). `hpaExpectedSnapshot` is only
        recomputed if `observedAt` changes.
      operationId: updateEmbryoObservation
      parameters: [ $ref: '#/components/parameters/OperatorId', $ref: '#/components/parameters/DeviceId', $ref: '#/components/parameters/IdempotencyKey' ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                observedAt: { type: string, format: date-time }
                outcome: { $ref: '#/components/schemas/EmbryoOutcome' }
                condition: { $ref: '#/components/schemas/Condition' }
                notes: { type: [string, 'null'] }
                correctionReason: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmbryoObservation' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Observations]
      summary: Soft-delete an observation
      operationId: deleteEmbryoObservation
      parameters: [ $ref: '#/components/parameters/OperatorId', $ref: '#/components/parameters/DeviceId', $ref: '#/components/parameters/IdempotencyKey' ]
      responses:
        '204': { description: Soft-deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    OperatorId:
      name: X-Operator-Id
      in: header
      required: true
      description: Who is doing this. Required on every write because there is no login (CON-01, FR-1105).
      schema: { type: string, format: uuid }
    DeviceId:
      name: X-Device-Id
      in: header
      required: true
      description: Stable per-device identifier generated on first use and kept in local storage.
      schema: { type: string, maxLength: 64 }
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      description: Stable key for one logical mutation. Replays return the original result.
      schema: { type: string, format: uuid }
...
  responses:
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflicts with current state — for example a duplicate normalised name, or a monotonic-survival violation without an override reason
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    EmbryoOutcome:
      type: string
      enum: [ALIVE, DEAD, DEGENERATED, NOT_OBSERVED]
      description: DEGENERATED is kept distinct from DEAD because the lab records the two separately.
...
    Condition:
      type: string
      enum: [NORMAL, ABNORMAL, UNDETERMINED]
...
    DueCheckpoint:
      type: object
      required: [injectionLotId, batchCode, lotNo, stageCode, stageLabel, dueAt, minutesLate, embryosRemaining]
      properties:
        injectionLotId: { type: string, format: uuid }
        batchId: { type: string, format: uuid }
        batchCode: { type: string }
        lotNo: { type: string }
        stageCode: { type: string }
        stageLabel: { type: string }
        stageOrder: { type: integer }
        dueAt: { type: string, format: date-time }
        minutesLate:
          type: integer
          description: Negative when the checkpoint is still in the future.
        embryosRemaining: { type: integer }
        operatorName: { type: [string, 'null'] }
    CheckpointEntry:
      type: object
      required: [injectionLotId, batchCode, lotNo, stage, activatedAt, expectedHpa, embryos]
      properties:
        injectionLotId: { type: string, format: uuid }
        batchCode: { type: string }
        lotNo: { type: string }
        stage: { $ref: '#/components/schemas/StageDefinition' }
        activatedAt: { type: string, format: date-time }
        expectedHpa: { type: number, examples: [2.5] }
        dueAt: { type: string, format: date-time }
        totalEmbryos: { type: integer }
        embryos:
          type: array
          items:
            type: object
            required: [embryoId, embryoCode, defaultCondition]
            properties:
              embryoId: { type: string, format: uuid }
              embryoCode: { type: string }
              wellPosition: { type: [string, 'null'] }
              defaultCondition:
                $ref: '#/components/schemas/Condition'
                description: Carried forward from the previous observation (FR-409).
              firstAbnormalStageLabel: { type: [string, 'null'] }
    EmbryoObservationInput:
      type: object
      required: [clientUuid, embryoId, stageCode, observedAt, outcome, condition]
      properties:
        clientUuid:
          type: string
          format: uuid
          description: Generated on the device before sending. The idempotency key (BR-18).
        embryoId: { type: string, format: uuid }
        stageCode: { type: string, examples: [stage_09_256C] }
        observedAt:
          type: string
          format: date-time
          description: Must fall between the lot's activation and now (BR-19).
        outcome: { $ref: '#/components/schemas/EmbryoOutcome' }
        condition: { $ref: '#/components/schemas/Condition' }
        notes: { type: [string, 'null'] }
        overrideReason:
          type: [string, 'null']
          maxLength: 500
          description: Required to record ALIVE for an embryo that already has an exit event (BR-07, FR-420).
    EmbryoObservationResult:
      type: object
      required: [clientUuid, status]
      properties:
        clientUuid: { type: string, format: uuid }
        id: { type: [string, 'null'], format: uuid }
        status: { $ref: '#/components/schemas/WriteStatus' }
        hpaActual: { type: [number, 'null'], examples: [2.6333] }
        hpaExpected: { type: [number, 'null'], examples: [2.5] }
        deviationH: { type: [number, 'null'], examples: [0.1333] }
        deviationLabel:
          type: [string, 'null']
          description: Preformatted for display per BR-23, so both platforms word it identically.
          examples: ['ช้ากว่าสากล 8 นาที']
        exitRecorded: { type: [boolean, 'null'] }
        error: { type: [object, 'null'], additionalProperties: true }
...